+++ matchbox-server/src/main/java/ch/ahdis/matchbox/CliContext.java
@@ -205,12 +205,6 @@ public class CliContext {
     return xVersion;
   }
 
-  private boolean devMode = false;
-
-  public boolean getDevMode() {
-	 return devMode;
-  }
-
   @Autowired
   public CliContext(Environment environment) {
     // get al list of all JsonProperty of cliContext with return values property
@@ -240,7 +234,6 @@ public class CliContext {
     this.autoInstallMissingIgs = environment.getProperty("matchbox.fhir.context.autoInstallMissingIgs", Boolean.class, false);
     this.extensions = Arrays.asList(environment.getProperty("matchbox.fhir.context.extensions", String[].class, new String[]{"any"}));
     this.xVersion = environment.getProperty("matchbox.fhir.context.xVersion", Boolean.class, false);
-    this.devMode = environment.getProperty("matchbox.fhir.context.devMode", Boolean.class, false);
   }
 
   public CliContext(CliContext other) {
@@ -262,7 +255,6 @@ public class CliContext {
     this.autoInstallMissingIgs = other.autoInstallMissingIgs;
     this.extensions = other.extensions;
     this.xVersion = other.xVersion;
-	 this.devMode = other.devMode;
   }
 
   @JsonProperty("ig")
@@ -692,7 +684,6 @@ public class CliContext {
         && xVersion == that.xVersion
         && httpReadOnly == that.httpReadOnly
         && autoInstallMissingIgs == that.autoInstallMissingIgs
-        && devMode == that.devMode
         && htmlInMarkdownCheck == that.htmlInMarkdownCheck
         && Objects.equals(extensions, that.extensions)
         && Objects.equals(txServer, that.txServer)
@@ -753,7 +744,6 @@ public class CliContext {
         locations,
         jurisdiction,
         onlyOneEngine,
-        devMode,
         xVersion);
     result = 31 * result + Arrays.hashCode(igsPreloaded);
     return result;
@@ -803,7 +793,6 @@ public class CliContext {
         ", xVersion=" + xVersion +
         ", httpReadOnly=" + httpReadOnly +
         ", autoInstallMissingIgs=" + autoInstallMissingIgs +
-        ", devMode=" + devMode +
         '}';
   }
 
@@ -849,7 +838,6 @@ public class CliContext {
 	addExtension(ext, "snomedCT", new StringType(this.snomedCT));
 	addExtension(ext, "fhirVersion", new StringType(this.fhirVersion));
 	addExtension(ext, "xVersion", new BooleanType(this.xVersion));
-	addExtension(ext, "devMode", new BooleanType(this.devMode));
 	addExtension(ext, "onlyOneEngine", new BooleanType(this.onlyOneEngine));
 	addExtension(ext, "ig", new StringType(this.ig));
 	// addExtension(ext, "questionnaireMode", new BooleanType(this.questionnaireMode));