+++ docs/changelog.md
@@ -1,6 +1,7 @@
 Unreleased
 
 - Upgrade hapifhir org.hl7.fhir.core to 6.4.4
+- Remove the `devMode` configuration parameter, it is now enabled when `httpReadOnly` is not [#315](https://github.com/ahdis/matchbox/issues/315)
 
 2024/11/25 Release 3.9.8